feat: re-write errorBtn to custom btn component on function
