+++ .eslintrc.cjs
@@ -40,7 +40,6 @@ module.exports = {
     '@typescript-eslint/no-var-requires': 'off',
     'react/jsx-uses-react': 'off',
     'react/react-in-jsx-scope': 'off',
-    'react/prefer-stateless-function': 'off',
   },
   settings: {
     react: {
