+++ src/components/UI/CustomBotton/CustomButton.tsx
@@ -1,0 +1,12 @@
+import ICustomButtonProps from './ICustomButtonProps';
+
+function CustomButton(props: ICustomButtonProps) {
+  const { disabled, children, onClick } = props;
+  return (
+    <button disabled={disabled} type="button" onClick={onClick}>
+      {children}
+    </button>
+  );
+}
+
+export default CustomButton;
