+++ src/components/UI/CustomBotton/ICustomButtonProps.ts
@@ -1,0 +1,8 @@
+import { MouseEvent } from 'react';
+
+interface ICustomButtonProps {
+  disabled: boolean;
+  children?: string | string[];
+  onClick: (event?: MouseEvent<HTMLElement>) => Promise<void> | void;
+}
+export default ICustomButtonProps;
